refactor(Palette): Paint 를 관리하는 Palette 정의
